♻️ File restructure of backend, improves readability
