+++ backend/src/main/kotlin/no/uib/echo/plugins/ContentNegotiation.kt
@@ -1,0 +1,17 @@
+package no.uib.echo.plugins
+
+import io.ktor.serialization.kotlinx.json.json
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
+import kotlinx.serialization.json.Json
+
+fun Application.configureContentNegotiation() {
+    install(ContentNegotiation) {
+        json(
+            Json {
+                ignoreUnknownKeys = true
+            }
+        )
+    }
+}
