+++ backend/src/main/kotlin/no/uib/echo/routes/UserRoutes.kt
@@ -1,0 +1,162 @@
+package no.uib.echo.routes
+
+import io.ktor.http.HttpStatusCode
+import io.ktor.server.application.Application
+import io.ktor.server.application.call
+import io.ktor.server.auth.authenticate
+import io.ktor.server.auth.jwt.JWTPrincipal
+import io.ktor.server.auth.principal
+import io.ktor.server.request.receive
+import io.ktor.server.response.respond
+import io.ktor.server.routing.Route
+import io.ktor.server.routing.get
+import io.ktor.server.routing.put
+import io.ktor.server.routing.routing
+import no.uib.echo.isEmailValid
+import no.uib.echo.schema.Degree
+import no.uib.echo.schema.StudentGroupMembership
+import no.uib.echo.schema.User
+import no.uib.echo.schema.UserJson
+import no.uib.echo.schema.bachelors
+import no.uib.echo.schema.masters
+import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
+import org.jetbrains.exposed.sql.StdOutSqlLogger
+import org.jetbrains.exposed.sql.addLogger
+import org.jetbrains.exposed.sql.insert
+import org.jetbrains.exposed.sql.lowerCase
+import org.jetbrains.exposed.sql.select
+import org.jetbrains.exposed.sql.transactions.transaction
+import org.jetbrains.exposed.sql.update
+
+fun Application.userRoutes() {
+    routing {
+        authenticate("auth-jwt") {
+            getUser()
+            putUser()
+        }
+    }
+}
+
+fun Route.getUser() {
+    get("/user") {
+        val email = call.principal<JWTPrincipal>()?.payload?.getClaim("email")?.asString()?.lowercase()
+
+        if (email == null) {
+            call.respond(HttpStatusCode.Unauthorized)
+            return@get
+        }
+
+        val user = transaction {
+            addLogger(StdOutSqlLogger)
+            User.select {
+                User.email eq email
+            }.firstOrNull()
+        }
+
+        if (user == null) {
+            call.respond(HttpStatusCode.NotFound, "User with email not found (email = $email).")
+            return@get
+        }
+
+        val memberships = transaction {
+            addLogger(StdOutSqlLogger)
+
+            StudentGroupMembership.select {
+                StudentGroupMembership.userEmail eq email
+            }.toList().map {
+                it[StudentGroupMembership.studentGroupName]
+            }
+        }
+
+        call.respond(
+            HttpStatusCode.OK,
+            UserJson(
+                user[User.email],
+                user[User.alternateEmail],
+                user[User.degreeYear],
+                Degree.valueOf(user[User.degree]),
+                memberships.ifEmpty { emptyList() }
+            )
+        )
+    }
+}
+
+fun Route.putUser() {
+    put("/user") {
+        val email = call.principal<JWTPrincipal>()?.payload?.getClaim("email")?.asString()?.lowercase()
+
+        if (email == null) {
+            call.respond(
+                HttpStatusCode.Unauthorized,
+                "Det har skjedd en feil. Vennligst prøv å logg inn og ut igjen."
+            )
+            return@put
+        }
+
+        try {
+            val user = call.receive<UserJson>()
+            val alternateEmail = user.alternateEmail?.lowercase()
+
+            if (alternateEmail != null) {
+                if (!isEmailValid(alternateEmail)) {
+                    call.respond(HttpStatusCode.BadRequest, "Vennligst skriv inn en gyldig e-post.")
+                    return@put
+                }
+            }
+
+            if (user.degreeYear !in 1..5) {
+                call.respond(HttpStatusCode.BadRequest, "Vennligst velg et gyldig trinn.")
+                return@put
+            }
+
+            val degreeMismatch = "Studieretning og årstrinn stemmer ikke overens."
+
+            if ((user.degree in bachelors && user.degreeYear !in 1..3) ||
+                (user.degree in masters && user.degreeYear !in 4..5) ||
+                (user.degree == Degree.ARMNINF && user.degreeYear != 1)
+            ) {
+                call.respond(HttpStatusCode.BadRequest, degreeMismatch)
+                return@put
+            }
+
+            val result = transaction {
+                addLogger(StdOutSqlLogger)
+                User.select {
+                    User.email.lowerCase() eq email
+                }.firstOrNull()
+            }
+
+            if (result == null) {
+                transaction {
+                    addLogger(StdOutSqlLogger)
+                    User.insert {
+                        it[User.email] = email
+                        it[User.alternateEmail] = alternateEmail
+                        it[degree] = user.degree.toString()
+                        it[degreeYear] = user.degreeYear
+                    }
+                }
+                call.respond(HttpStatusCode.OK, "User created with email = $email")
+                return@put
+            }
+
+            transaction {
+                addLogger(StdOutSqlLogger)
+                User.update({
+                    User.email.lowerCase() eq email
+                }) {
+                    it[User.alternateEmail] = alternateEmail
+                    it[degree] = user.degree.toString()
+                    it[degreeYear] = user.degreeYear
+                }
+            }
+            call.respond(
+                HttpStatusCode.OK,
+                "User updated with email = $email, alternateEmail = $alternateEmail, degree = ${user.degree}, degreeYear = ${user.degreeYear}"
+            )
+        } catch (e: Exception) {
+            call.respond(HttpStatusCode.InternalServerError)
+            e.printStackTrace()
+        }
+    }
+}
